[TASK] Split TtContentOverridesAggregate
