+++ Classes/Aggregate/AbstractOverridesAggregate.php
@@ -25,50 +25,20 @@ namespace CPSIT\MaskExport\Aggregate;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
-use CPSIT\MaskExport\CodeGenerator\HtmlCodeGenerator;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
-
 /**
  * @package mask
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  */
-abstract class AbstractOverridesAggregate extends AbstractAggregate implements LanguageAwareInterface, PhpAwareInterface, PlainTextFileAwareInterface, SqlAwareInterface
+abstract class AbstractOverridesAggregate extends AbstractAggregate implements PhpAwareInterface, SqlAwareInterface
 {
     use PhpAwareTrait;
-    use PlainTextFileAwareTrait;
     use SqlAwareTrait;
     use TcaAwareTrait;
-
-    /**
-     * @var HtmlCodeGenerator
-     */
-    protected $htmlCodeGenerator;
 
     /**
      * @var string
      */
     protected $tcaOverridesFilePath = 'Configuration/TCA/Overrides/';
-
-    /**
-     * @var string
-     */
-    protected $templatesFilePath = 'Resources/Private/Templates/';
-
-    /**
-     * @var string
-     */
-    protected $typoScriptFilePath = 'Configuration/TypoScript/';
-
-    /**
-     * @param array $maskConfiguration
-     * @param HtmlCodeGenerator $htmlCodeGenerator
-     */
-    public function __construct(array $maskConfiguration, HtmlCodeGenerator $htmlCodeGenerator = null)
-    {
-        $this->htmlCodeGenerator = (null !== $htmlCodeGenerator) ? $htmlCodeGenerator : GeneralUtility::makeInstance(HtmlCodeGenerator::class);
-
-        parent::__construct($maskConfiguration);
-    }
 
     /**
      * @param array $tableConfiguration
@@ -94,20 +64,5 @@ abstract class AbstractOverridesAggregate extends AbstractAggregate implements L
 
 EOS
         );
-    }
-
-    /**
-     * @param array $element
-     */
-    protected function addFluidTemplate(array $element)
-    {
-        $key = $element['key'];
-        $html = $this->htmlCodeGenerator->generateHtml($key);
-        if (!empty($html)) {
-            $this->addPlainTextFile(
-                $this->templatesFilePath . GeneralUtility::underscoredToUpperCamelCase($this->table) . '/'. $key . '.html',
-                $html
-            );
-        }
     }
 }
